# api.yml
openapi: '3.0.4'
info:
  title: Sports Facilities Management System
  description: |-
    Esta es la definición de la API que se va a usar dentro del sistema de gestión de centros deportivos para poder hacer uso de todas las funciones que ofrece. Estas pueden ser separadas en función del tipo de usuario final que va a poder acceder a ellas:
    - Administradores: gestión de instalaciones, dispositivos y usuarios.
    - Clientes: gestión del perfil y reserva de instalaciones.
  contact:
    email: pabloquesadarojo@gmail.com
  version: '1.3.0'
servers:
  - url: https://api.server.test/
security:
  - BearerAuth: []
...
paths:
...
  /complexes:
    get:
      tags:
        - complexes
      summary: Obtiene una lista de complejos deportivos del sistema.
      description: Obtiene una lista de complejos deportivos del sistema que puede estar filtrada en función de una serie de parámetros.
      operationId: getComplexes
      x-security-roles:
        - ADMIN
        - USER
      responses:
        '200':
          description: Se devuelve correctamente la lista de complejos deportivos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ComplexInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - complexes
      summary: Crea un nuevo complejo deportivo en el sistema.
      description: Crea un nuevo complejo deportivo en el sistema, asignándole un identificador único y completando la información asociada a este.
      operationId: createComplex
      x-security-roles:
        - SUPERADMIN
      requestBody:
        description: Objeto que contiene la información del complejo deportivo a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplexInfo'
        required: true
      responses:
        '200':
          description: El complejo deportivo se ha creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplexInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complexes/{complexId}/devices:
    parameters:
      - $ref: '#/components/parameters/complexId'
    get:
      tags:
        - devices
      summary: Obtiene una lista de dispositivos del sistema.
      description: Obtiene una lista de dispositivos del sistema que puede estar filtrada en función de una serie de parámetros.
      operationId: getDevices
      x-security-roles:
        - ADMIN
      parameters:
        - name: type
          in: query
          description: Tipo de dispositivo a filtrar.
          schema:
            $ref: '#/components/schemas/DeviceType'
        - name: status
          in: query
          description: Estado de funcionamiento del dispositivo a filtrar.
          schema:
            $ref: '#/components/schemas/DeviceStatus'
      responses:
        '200':
          description: Se devuelve correctamente la lista de dispositivos.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - devices
      summary: Crea un nuevo dispositivo en el sistema.
      description: Crea un nuevo dispositivo en el sistema, asignándole un identificador único y un tipo.
      operationId: createDevice
      x-security-roles:
        - ADMIN
      requestBody:
        description: Objeto que contiene la información del dispositivo a crear.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInfo'
        required: true
      responses:
        '201':
          description: El dispositivo se ha creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /complexes/{complexId}/devices/{deviceId}:
    parameters:
      - $ref: '#/components/parameters/complexId'
      - $ref: '#/components/parameters/deviceId'
    get:
      tags:
        - devices
      summary: Obtiene la información de un dispositivo.
      description: Obtiene la información de un dispositivo concreto.
      operationId: getDeviceInfo
      x-security-roles:
        - ADMIN
      responses:
        '200':
          description: Se devuelve correctamente la información del dispositivo pedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - devices
      summary: Actualiza la información de un dispositivo.
      description: Actualiza la información de un dispositivo concreto.
      operationId: updateDeviceInfo
      x-security-roles:
        - ADMIN
      requestBody:
        description: Objeto que contiene la información actualizada del dispositivo.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInfo'
        required: true
      responses:
        '200':
          description: La información del dispositivo se ha actualizado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceInfo'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - devices
      summary: Elimina un dispositivo del sistema.
      description: Elimina un dispositivo del sistema, borrando toda su información asociada.
      operationId: deleteDevice
      x-security-roles:
        - ADMIN
      responses:
        '204':
          description: El dispositivo se ha eliminado correctamente.
          content:
            application/json: {}
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /complexes/{complexId}/devices/{deviceId}/telemetry:
    parameters:
      - $ref: '#/components/parameters/complexId'
      - $ref: '#/components/parameters/deviceId'
    get:
      tags:
        - devices
      summary: Obtiene la telemetría de un dispositivo.
      description: Obtiene la telemetría de un dispositivo.
      operationId: getDeviceTelemetry
      x-security-roles:
        - ADMIN
      responses:
        '200':
          description: Se devuelve correctamente la telemetría del dispositivo pedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTelemetry'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - devices
      summary: Carga la telemetría de un dispositivo en el sistema.
      description: Carga la telemetría de un dispositivo en el sistema, actualizando el estado interno y llevando a cabo un análisis de los datos.
      operationId: addDeviceTelemetry
      security:
        - DeviceAuth: []
      x-security-roles:
        - ADMIN
        - DEVICE
      requestBody:
        description: Objeto que contiene la información para identificar el dispositivo que envía los datos y el valor de la telemetría.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTelemetry'
        required: true
      responses:
        '201':
          description: La telemetría se ha cargado correctamente en el sistema.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTelemetry'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /complexes/{complexId}/devices/{deviceId}/status:
    parameters:
      - $ref: '#/components/parameters/complexId'
      - $ref: '#/components/parameters/deviceId'
    get:
      tags:
        - devices
      summary: Obtiene el estado de funcionamiento de un dispositivo.
      description: Obtiene el estado de funcionamiento de un dispositivo.
      operationId: getDeviceStatus
      security:
        - BearerAuth: []
        - DeviceAuth: []
      x-security-roles:
        - ADMIN
        - DEVICE
      responses:
        '200':
          description: Se devuelve correctamente el estado del dispositivo pedido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se ha encontrado el dispositivo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - devices
      summary: Carga el estado de funcionamiento de un dispositivo en el sistema.
      description: Carga el estado de funcionamiento de un dispositivo en el sistema, actualizando el estado interno y llevando a cabo un análisis de los datos.
      operationId: addDeviceStatus
      x-security-roles:
        - ADMIN
        - DEVICE
      requestBody:
        description: Objeto que contiene la información para identificar el dispositivo que envía los datos y el valor del estado.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceStatus'
        required: true
      responses:
        '201':
          description: El estado se ha cargado correctamente en el sistema.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatus'
        '400':
          description: Petición incorrecta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No se ha podido autenticar al usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Sin permisos para realizar la operación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          description: Código de estado HTTP.
        message:
          type: string
          description: Mensaje descriptivo del error.
        error:
          type: string
          description: Tipo de error (ej. Bad Request, Unauthorized).
...
    WeatherData:
      type: object
      properties:
        temperature:
          type: number
          description: Temperatura actual.
        precip_intensity_prev:
          type: number
          description: Precipitación registrada en la hora previa.
        precip_intensity_curr:
          type: number
          description: Precipitación actual.
        precip_probability_curr:
          type: integer
          description: Probabilidad de precipitación actual.
        precip_probability_next:
          type: integer
          description: Probabilidad de precipitación para la próxima hora.
      description: Datos meteorológicos obtenidos de la API de OpenMeteo.
    DeviceType:
      type: string
      enum:
        - presence
        - rain
      description: Tipo de dispositivo IoT.
    DeviceStatus:
      type: string
      enum:
        - normal
        - off
        - battery
        - error
      default: normal
      description: Estado del dispositivo IoT.
    DeviceInfo:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único del dispositivo IoT.
        type:
          $ref: '#/components/schemas/DeviceType'
      required:
        - type
    DeviceTelemetry:
      type: integer
...
    TimeInfo:
      type: object
      properties:
        timeIni:
          type: string
          format: time
        timeEnd:
          type: string
          format: time
    ComplexInfo:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        time:
          $ref: '#/components/schemas/TimeInfo'
        locLongitude:
          type: number
          format: double
        locLatitude:
          type: number
          format: double
        weather:
          $ref: '#/components/schemas/WeatherData'
...
  parameters:
    complexId:
      name: complexId
      in: path
      required: true
      description: Identificador del sistema.
      schema:
        type: integer
    deviceId:
      name: deviceId
      in: path
      required: true
      description: Identificador del dispositivo.
      schema:
        type: integer
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceAuth:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Token de autenticación del dispositivo, utilizado para autenticar las peticiones de los dispositivos IoT.